hide branches for sole proprietors
